requirements.txt
torch
torchvision
tqdm
mediapipe
Pillow
diffusers==0.25.0
transformers==4.28.1
hf_transfer
fire
huggingface_hub==0.24.0
imageio
imageio-ffmpeg
PyYAML
safetensors
wandb
gradio
plyfile
xformers
tensorboard
tensorboard-data-server
wandb
jaxtyping
einops
omegaconf
hydra-core>=1.3
kornia
torchtyping
gsplat
mpl_scatter_density
labtasker[plugins]
envidia
black
isort
pre-commit
